Query hfp by time. Still have some cache things to figure out...
